Update comment extractor
