+++ package.json
@@ -8,6 +8,7 @@
   },
   "dependencies": {
     "async": "^2.0.0-rc.4",
+    "pandoc-filter": "^0.1.6",
     "pdc": "^0.2.2",
     "ramda": "^0.21.0"
   },
